requirements.txt
transformers>=4.40.0
torch>=2.0.0
datasets>=2.18.0
tokenizers>=0.19.0
accelerate>=0.28.0
sentencepiece>=0.2.0
protobuf>=4.21.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
requests>=2.25.0
tqdm>=4.62.0
# Dependencias para generación de imágenes
diffusers>=0.21.0,<0.30.0
pillow>=9.0.0
safetensors>=0.3.0
# xformers>=0.0.16  # Comentado: causa problemas de compatibilidad
